Refactor: 레이아웃 변경
현재 person routes에 +page 에 들어있던 input box를 layout으로 변경
person ID를 통해 [personID] 라우팅을 추가하기 위함
input box 상단 header에 가려지는 버그 수정
